Prevent runner to be stealth with long logs.
If the log is longer than maxTraceOutputSize, the runner will be paused
and waiting forever. This ia because the reader stop reading the pipe,
and the pipes has a size limit.

I have already proposed a patch in the past for this. The patch was
adapted, but unfortunately, an important "break" was not removed in the
adaptation.

This patch remove the unwanted break, allowing the for loop to continue
to run and flushing the pipe buffer.