Fix files configured in "cache" never being archived
When `archiveFiles`, part of BashShell, is passed
the configuration for the cache (coming from
`.gitlab-ci.yml`, it is expected to be of type
`map[string]interface{}`. The actual type coming in
from the YAML parser is `map[interface{}]interface{}`.
As a result, the type assertion never holds and thus
files are never archived during build.