Add a more verbose description of docker-ssh
I was mistakenly under the impression that with `docker-ssh` it would
ssh into a remote machine and then run docker on that machine. It took
me a while of debugging and looking through the source to finally
realize that it was running Docker locally, and then SSHing into that
container.

I figured I am not the only one to be confused, so I have added a
little more explanation to the docs.