Reconcile VirtualBox status constants with VBoxManage output values
This replaces the existing set of constants for the VirtualBox
StatusType values with constants based on the actual values that may be
returned when calling "VBoxManage showvminfo". The function
IsStatusOnlineOrTransient() has also been added to simplify checking
whether a status value indicates that the VM is running or in a
transitional state (but not offline), and the VirtualBox executor
Prepare() function has been updated to use it to more accurately check
whether a VM is offline before starting it.