Reimplemented invalid VirtualBox VM check in createVM()
Checks for "Invalid" VM statuses were previously removed from
executor_virtualbox.go since "Invalid" was never a status returned from
the VirtualBox Status() function. The test in createVM() has been
re-added, instead testing for an error result, in case an invalid VM
exists that causes VBoxManage to return an error so we can still attempt
to unregister it.

The test in Prepare() has not been restored, as the VM name string it
tests is not initialized at that point. Also, the subsequent Exist()
checks return false if VBoxManage returns an error, causing createVM()
to eventually be called anyway.