Add support for linked cloning of VirtualBox VMs
This commit modifies the VirtualBox runner to create linked clones
instead of full clones of the base VM when possible. The runner checks
for any current VM snapshot and attempts to create a linked clone from
it. If no current snapshot is found and snapshots are not disabled, the
runner will attempt to create a new snapshot first. If this fails or if
snapshots are disabled, the runner will fall back on creating a full
clone of the current VM state.

The snapshot from which a linked clone is created can also be explicitly
set using the "base_snapshot" configuration option. If the snapshot
doesn't exist and snapshots are not disabled, the runner will attempt
to create one with the same name before falling back on creating a full
VM clone.