Use .xz for prebuilt docker images
This allows to reduce the binary size by about 20MB. The change also makes the archive to be decompressed by Docker Engine, not by runner reducing the CPU and Network usage when starting a new build.

This also introduces compatibility tests up to docker 1.8 to make sure that all features do work as expected. Possibly we should make 1.8 a minimal requirement for GitLab Runner.