Suppress all but the first warning of a given type when extracting a ZIP file
Operations like lchown() and chmod() are called once for each entry in a ZIP
archive. If the operation fails for the first entry, it's very likely to fail
for every one in the archive. The errors are non-fatal, but lead to annoying
build log spam.

This commit changes the output so failure messages are only shown the first
time an operation files.